pneumonia-backend/requirements.txt
fastapi
uvicorn
fastai==2.7.12
torch
torchvision
pillow
python-multipart